added VehicleDetails and fixed a bunch of bugs
