+++ src/BookingCost.js
@@ -8,16 +8,13 @@ export const calculateBookingCost = (booking, bookingJourneys) => {
 
   if (booking.bookingType === BookingType.PER_DAY) {
 	bookingCost = calculateBookingDurationInDays(booking) * COST_PER_DAY;
-  }
-  else {
-    if (Number.isNaN(calculateDistanceTravelledInKm(booking, findLastJourney(booking, bookingJourneys)))) {
-      bookingCost = calculateDistanceTravelledInKm(booking, findLastJourney(booking, bookingJourneys));
-	}
-    else {
-	  bookingCost = calculateDistanceTravelledInKm(booking, findLastJourney(booking, bookingJourneys)) * COST_PER_KM;
+  } else {
+	if (Number.isNaN(calculateDistanceTravelledInKm(booking, findLastOdometerReading(bookingJourneys)))) {
+	  bookingCost = calculateDistanceTravelledInKm(booking, findLastOdometerReading(bookingJourneys));
+	} else {
+	  bookingCost = calculateDistanceTravelledInKm(booking, findLastOdometerReading(bookingJourneys)) * COST_PER_KM;
 	}
   }
-
   return bookingCost;
 };
 
@@ -31,9 +28,9 @@ export const calculateRevenueRecorded = (bookings, journeys) => {
   }, 0)).toFixed(2);
 };
 
-const calculateDistanceTravelledInKm = (booking, lastJourney) => {
-  if (lastJourney) {
-	const endOdometer = lastJourney.journeyEndOdometerReading;
+const calculateDistanceTravelledInKm = (booking, lastOdometerReading) => {
+  if (lastOdometerReading) {
+	const endOdometer = lastOdometerReading;
 	const startOdometer = booking.startOdometer;
 
 	return endOdometer - startOdometer;
@@ -48,24 +45,11 @@ const calculateBookingDurationInDays = booking => {
   return bookingEndDate.diff(bookingStartDate, 'days');
 };
 
-const findLastJourney = (booking, bookingJourneys) => {
-  if (bookingJourneys.length) {
-	const bookingEndDate = moment(booking.endDate);
-	let lastJourneyIndex = 0;
-	let minimumDateDiff = bookingEndDate.diff(moment(bookingJourneys[lastJourneyIndex].journeyEndedAt), 'days');
-	let currentDiff = minimumDateDiff;
-
-	bookingJourneys.forEach((journey, index) => {
-	  currentDiff = bookingEndDate.diff(moment(journey.journeyEndedAt), 'days');
-	  if (currentDiff < minimumDateDiff) {
-		currentDiff = minimumDateDiff;
-		lastJourneyIndex = index;
-	  }
-	});
-
-	return bookingJourneys[lastJourneyIndex];
+const findLastOdometerReading = bookingJourneys => {
+  if (bookingJourneys && bookingJourneys.length) {
+	return bookingJourneys
+	  .map(j => j.journeyEndOdometerReading)
+	  .reduce((a, b) => Math.max(a, b))
   }
-
-  return null;
+  return 0;
 };
-
